# Move stage 2 parser state out of simdjson.h