This creates a way for implementations to allocate and store their own internal state. It only moves stage 2 state over, however; some of our benchmarking statistics tools rely on structural_indexes, so I figure on moving that stuff in a separate PR.

How it works:

- We reserve a 24-byte region inside the dom::parser at `parser._implementation_state`. (We will widen this as we accomodate future implementations.) This is the region that implementations will overwrite and manipulate with their own data. It is initialized with zeroes.
- `T& parser.implementation_state<T>()` reinterpret_casts that 24-byte region to the type of your choice. This is what implementations will use.
- parser.allocate() calls `active_implementation->allocate()` at allocation time, allowing the implementation to perform whatever allocations it needs.
- The stage2 code defines `parser_state` in parser_state.h, and calls `parser.implementation_state<parser_state>()` to retrieve it. This is meant to be a zero-cost operation, as all it should be doing is returning the pointer with different type information.
- current_loc, containing_scope, and ret_address (along with their types) are moved 

## Rationale

My primary rationale is actually that I want a place to save in-progress state from stage 1 and 2 when streaming, and some of the stuff I want to save is simd8<uint8_t>, which is different for each implementation.

Secondary, however, is I want to be able to experiment with different backends without breaking the primary one. This is an important step towards being able to *check in* an experimental implementation that does things differently, so that we can keep it up to date but not actually use it in production. Pass `SIMDJSON_FORCE_IMPLEMENTATION=hsawell-experiment`.

## Performance

On Skylake, Skylake-X and my Kady Lake R machine, this makes no performance difference on twitter. However, it does increase instruction count by about 1 instruction per 20 structural characters. I think it's worth that cost, given no actual difference and that we can begin to freely change how implementations store data internally.

### Performance Learnings

A couple of things became clear here about some of the fiddliness of stage 2:

* If `current_loc` is not the very first thing in the parser, stage 2 loses around 5% throughput. I suspect not having the direct pointer to current_loc prevents the compiler from using a more efficient single mov instruction that includes the increment in `tape[current_loc++] = ...`. Just a suspicion based on where I have seen this issue crop up, however.
* If `containing_scope` is a simple pointer instead of a `unique_ptr`, stage 2 loses around 5% throughput. This is utterly baffling to me.