@jkeiser 2020-05-23
It passes the tests and doesn't hurt perf on any machine I've tried. However, it seems some jerk pushed a conflicting change that writes the start/end document in different places, and *that* patch combined with this one causes a 2% regression. Not sure why yet.