@jkeiser 2020-05-23
... and it appears the reason was that when I merged it the first time I did it wrong. Pushed a version rebased against master with performance to boot.