@jkeiser 2020-05-23
@lemire I [found the fix](https://github.com/simdjson/simdjson/pull/902/commits/04cd8fab498dd9ac962a5275f662c7c6d9107ca3) for the performance regression after rebasing, and it's a doozy. Basically, instead of just incrementing current_loc when we first encounter the start of an array or object (which we'll write out for real later), my change writes a *zero* to it. This restores performance.

Pure rampant speculation: the compiler notices that between every goto we append to the tape, and does an optimization. Less rampant speculation: the compiler is at a sort of "random valley" where any little change bumps it into a random place performance-wise.