@jkeiser 2020-05-23
Final push. At this point I've concluded that we're in a "valley of randomness" where any change pushes stage 2 either + or -5%. I found that things were slower on *my* machine after rebasing against master, but OK on skylake. I found a fix that made things faster on my machine, which made it OK on my machine, and slowed down skylake.

So I'm leaving it as is and concluding this doesn't hurt performance at all.