@jkeiser 2020-05-26
@lemire here's a more complete accounting. I'm seeing an across-the-board degradation on capacity allocation on Windows. This might well have to do with inlining or with putting allocations close together (or both), but its effect is limited to the truly huge allocations. I think that can be recovered pretty easily, because the truly huge allocations don't have to be done by stage 1 (yet). I'll fix that, but this gives us food for thought ...

| Preallocate | Master Allocation Throughput (GB/s) | PR Allocation Throughput (GB/s) |
|---|---|---|
| default | 39.4697 | 30.5080 |
| allocate(1, 1) | 38.9824 | 30.2160 |
| allocate(1, DEFAULT_MAX_DEPTH) | 40.2239 | 32.7210 |
| allocate(data.size(), DEFAULT_MAX_DEPTH) | 0 | 0 |
| allocate(data.size(), 1) | 0 | 0 |